separate test_write_out_html()
