+++ tests/test_app.py
@@ -1,12 +1,10 @@
 """ test app.py functions """
-import datetime
 import os
 import typing
 import pytest
 from qsl_pdf_publisher.log_file_utils.log_file import LogFile
 from qsl_pdf_publisher.log_file_utils.get_log_file_list import get_log_file_list
-from qsl_pdf_publisher.app import get_log_file_list, write_out_html, write_out_pdf
-from qsl_pdf_publisher.qso import Qso
+from qsl_pdf_publisher.app import get_log_file_list, write_out_pdf
 
 
 def test_get_log_file_list(adif_file_list: typing.Tuple[LogFile, ...], tmpdir) -> None:
@@ -32,51 +30,6 @@ def adif_file_list(tmpdir) -> typing.Tuple[LogFile, ...]:
 
     for basename in basename_list:
         os.remove(os.path.join(tmpdir, basename))
-
-
-def test_write_out_html(qso_log: typing.Tuple[Qso, ...], html_file_path: str):
-    """ test write_out_html() """
-    write_out_html(qso_log, html_file_path)
-
-    assert os.path.isfile(html_file_path)
-
-    with open(html_file_path, 'r', encoding='utf-8') as f:
-        html = f.read()
-
-    assert 'A1B2C3' in html
-
-
-@pytest.fixture
-def qso_log() -> typing.Tuple[Qso, ...]:
-    """ fixture returns list of qso """
-
-    yield tuple([Qso(
-        callsign='A1B2C3',
-        transfer_callsign='A1B2C3',
-        datetime=datetime.datetime(2001, 12, 1, 23, 43),
-        rst_sent='599',
-        band='40m',
-        mode='CW',
-        comment='FB QSO TNX',
-        frequency='7',
-        my_antenna='Super Ultra Great Wonderful DP',
-        my_qth='Top of Mt. Fuji (JCC#1807)',
-        my_cq_zone='25',
-        my_dxcc='339',
-        my_gridsquare='PM95',
-        my_iota='',
-        my_rig='FTDX101MP',
-    )])
-
-
-@pytest.fixture
-def html_file_path(tmpdir) -> str:
-    """ fixture returns path to html file """
-    path = os.path.join(tmpdir, 'unittest_out.html')
-
-    yield os.path.join(path)
-
-    os.remove(path)
 
 
 def test_write_out_pdf(html_file_path_to_write_out_pdf: str, css_file_path: str, pdf_file_path: str):
